Optimize loop increment to save gas
Addresses the valid bug report on gas optimization in for-loop increments.

Issue:
The use of post increments (i++) in for loops in the StakingRewards.sol contract resulted in unnecessary gas costs.

Resolution:
Switched to using pre-increment (++i) in the affected for loops to save around 600 gas units during contract deployment. This change keeps the function logic the same while optimizing for gas.